✨ Refactor ProfilePage and extract ProfileForm component
- Moved the profile form logic into a new `ProfileForm` component for better separation of concerns.
- Updated `ProfilePage` to handle session management and redirection using `getServerSession`.
- Improved loading state handling by removing unnecessary loading indicators from the main page component.
